# Conflicts:
#	erp-server/erp-server-mrp/src/main/java/com/erp/server/mrp/service/impl/ReplenishmentSuggestionServiceImpl.java

+++ erp-sdk/erp-sdk-oms-mercado/src/main/java/com/sdk/oms/mercado/dto/mercado/order/BuyerBean.java
@@ -12,6 +12,6 @@ public class BuyerBean {
      */
 
     @JsonProperty("id")
-    private int fid;
+    private long fid;
 
 }
